+++ scripts/dev-tools-install.sh
@@ -9,7 +9,9 @@ source "${STACK_ROOT}/includes/function.inc.sh"
 title "서버 필수 유틸과 점검툴들을 설치합니다."
 
 # stack.conf 에서 선언한 패키지 설치
-if [ "$OS" = "rocky8" ]; then
+if [ "$OS" = "rocky9" ]; then
+  PACKAGES=${DEV_PACKAGES_R9}
+elif [ "$OS" = "rocky8" ]; then
   PACKAGES=${DEV_PACKAGES_R8}
 else
   PACKAGES=${DEV_PACKAGES}